feat: we can lazy load the autocapture
