+++ packages/browser/playwright/mocked/utils/posthog-playwright-test-base.ts
@@ -14,6 +14,8 @@ const lazyLoadedJSFiles = [
     'tracing-headers',
     'web-vitals',
     'dead-clicks-autocapture',
+    'autocapture',
+    'heatmaps',
 ]
 
 export type WindowWithPostHog = typeof globalThis & {
