+++ packages/browser/rollup.config.mjs
@@ -182,6 +182,12 @@ const plugins = (es5, noExternal) => [
                               // set as part of lazy-loading (doesn't start with _ BUT be abundantly cautious)
                               'loadExternalDependency',
 
+                              // called across main bundle / lazy bundle boundary for autocapture
+                              '_captureEvent',
+
+                              // called across main bundle / lazy bundle boundary for heatmaps
+                              '_onClick',
+
                               // part of the public API (none start with _ so are not mangled anyway BUT be abundantly cautious)
                               'capture',
                               'identify',
