+++ packages/browser/src/entrypoints/dead-clicks-autocapture.ts
@@ -1,9 +1,9 @@
 import { assignableWindow, LazyLoadedDeadClicksAutocaptureInterface } from '../utils/globals'
 import { PostHog } from '../posthog-core'
 import { isNull, isNumber, isUndefined } from '@posthog/core'
-import { autocaptureCompatibleElements, getEventTarget } from '../autocapture-utils'
+import { autocaptureCompatibleElements, getEventTarget } from '../utils/dom-event-utils'
 import { DeadClickCandidate, DeadClicksAutoCaptureConfig, Properties } from '../types'
-import { autocapturePropertiesForElement } from '../autocapture'
+import { autocapturePropertiesForElement } from '../extensions/autocapture/external'
 import { isElementInToolbar, isElementNode, isTag } from '../utils/element-utils'
 import { getNativeMutationObserverImplementation } from '../utils/prototype-utils'
 import { addEventListener } from '../utils'
