+++ packages/browser/src/posthog-core.ts
@@ -1,4 +1,4 @@
-import { Autocapture } from './autocapture'
+import { Autocapture } from './extensions/autocapture'
 import Config from './config'
 import { ConsentManager, ConsentStatus } from './consent'
 import {
@@ -20,7 +20,7 @@ import { SentryIntegration, sentryIntegration, SentryIntegrationOptions } from '
 import { Toolbar } from './extensions/toolbar'
 import { TracingHeaders } from './extensions/tracing-headers'
 import { WebVitalsAutocapture } from './extensions/web-vitals'
-import { Heatmaps } from './heatmaps'
+import { Heatmaps } from './extensions/heatmaps'
 import { PageViewManager } from './page-view'
 import { PostHogExceptions } from './posthog-exceptions'
 import { PostHogFeatureFlags } from './posthog-featureflags'
