+++ packages/browser/terser-mangled-names.json
@@ -42,7 +42,6 @@
         "_canvasRecording",
         "_capture",
         "_captureDeadClick",
-        "_captureEvent",
         "_captureExtensionExceptions",
         "_captureInitialPageview",
         "_capturePageview",
@@ -104,6 +103,7 @@
         "_evaluateAndDisplayTours",
         "_eventCollector",
         "_eventEmitter",
+        "_eventQueue",
         "_eventToSurveys",
         "_eventTriggerMatching",
         "_eventTriggerStatus",
@@ -212,6 +212,7 @@
         "_lastPathname",
         "_lastSelectionChanged",
         "_lazyLoadAndStart",
+        "_lazyLoadedAutocapture",
         "_lazyLoadedDeadClicksAutocapture",
         "_lazyLoadedSessionRecording",
         "_linkedFlagMatching",
@@ -301,6 +302,7 @@
         "_productTourManager",
         "_queue",
         "_queuedRRWebEvents",
+        "_rageclicks",
         "_random",
         "_rateLimiter",
         "_receivedFlags",
